refactor app layout and  error handler + remove obsolete code
