+++ src/actions.js
@@ -9,7 +9,4 @@ export const SET_ADVANCED_MODE = "asab/setAdvancedMode";
 export const CHANGE_CONFIG = "asab/changeConfig";
 export const SET_DEV_CONFIG = "asab/setDevConfig";
 
-export const SET_SIDEBAR = "asab/setSidebar";
-export const CHANGE_SIDEBAR_SIZE = "asab/changeSidebarSize";
-export const SET_SMALL_SIDEBAR = "asab/setSmallSidebar";
 export const SET_HIDDEN_ITEMS_SIDEBAR = "asab/setHiddenItemsSidebar";
