+++ src/styles/components/breadcrumbs.scss
@@ -1,0 +1,9 @@
+.breadcrumbs {
+	background-color: $breadcrumb-bg;
+}
+
+.breadcrumb {
+	margin: 0 !important;
+	padding: 0;
+	background-color: inherit;
+}
