+++ src/styles/components/index.scss
@@ -1,1 +1,6 @@
+@import "./breadcrumbs.scss";
 @import "./card.scss";
+@import "./sidebar.scss";
+@import "./main.scss";
+@import "./header.scss";
+@import "./error-handler.scss";
