+++ src/styles/components/main.scss
@@ -1,0 +1,26 @@
+// might be removed after refactoring
+.main {
+	margin-top: 55px;
+}
+
+@media screen and (max-width: 768px) {
+	.main, .main-with-sidebar, .main-with-sidebar-minimized {
+		margin-left: 0px !important;
+	}
+
+	// remove stop propagation container later
+	.stop-propagation-container {
+		height: calc(100% + 50px);
+		width: 100%;
+		position: absolute;
+		top: 55px;
+		background-color: rgba(0,0,0,0.5);
+		overflow-y: hidden;
+		z-index: 999;
+	}
+}
+
+// Add a vertical space for a standard screen layout
+main > .container-fluid, main > .container {
+	margin-top: 1rem; // basically mt-3
+}
