+++ src/styles/components/sidebar.scss
@@ -1,0 +1,159 @@
+.app-sidebar {
+	margin-left: 0;
+	width: 100px;
+	transition: margin-left 0.25s ease-out, width .2s ease-out;
+	height: calc(100vh - 55px);
+	position: fixed;
+	z-index: 1019;
+	display: flex;
+	padding: 0;
+	color: #fff;
+	background-color: $sidebar-bg;
+	flex: 0 0 100px;
+	order: -1;
+
+	// This should be refactored for on hover
+	.sidebar-item-text, .sidebar-chevron, .sidebar-chevron-open {
+		display: none !important;
+	}
+
+	.sidebar-dropdown-open {
+		background-color: rgba(0,0,0,.2);
+	}
+	.sidebar-dropdown {
+		.sidebar-chevron, .sidebar-chevron-open {
+			font-size: 0.5rem;
+			margin: 0.3rem;
+			float: right;
+			.sidebar-item-icon {
+				margin: 0;
+				font-weight: 700;
+				transition: transform .1s ease-in-out;
+			}
+		}
+		.sidebar-chevron-open {
+			.sidebar-item-icon {
+				transform: rotate(-90deg);
+			}
+		}
+	}
+	.sidebar-nav {
+		position: relative;
+		flex: 1;
+		width: 100px;
+		overflow: hidden;
+		touch-action: auto;
+		overflow-anchor: none;
+		height: 100%;
+		display: flex;
+		flex-direction: column;
+		justify-content: space-between;
+		.nav {
+			width: 100px;
+			display: flex;
+			flex-direction: column;
+			flex-wrap: nowrap;
+			padding: 0;
+			margin: 0 auto;
+			overflow-y: auto;
+			transition: width .25s ease-in-out;
+			.nav-item {
+				position: relative;
+				margin: 0;
+				.nav-link {
+					padding: 0.75rem 1rem;
+					color: #fff;
+					.inner-link-container {
+						display: flex;
+						justify-content: space-between;
+						align-items: center;
+						flex-wrap: nowrap;
+						overflow-x: hidden;
+					}
+					.sidebar-item-icon {
+						display: inline-block;
+						width: 1rem;
+						margin: 0 0.5rem 0 0;
+						color: #73818f;
+						text-align: center;
+					}
+				}
+
+				.nav-link:hover {
+					cursor: pointer;
+					color: #fff;
+					background: #20a8d8;
+					.sidebar-item-icon {
+						color: #fff;
+					}
+				}
+
+				.active {
+					color: #fff;
+					background-color: #3a4248;
+					.sidebar-item-icon {
+						color: #20a8d8;
+					}
+					.sidebar-item-icon {
+						color: #20a8d8;
+					}
+				}
+
+				.nav-children {
+					transition: width 0s;
+					overflow: hidden;
+				}
+			}
+		}
+	}
+	.sidebar-bottom {
+		height: 50px;
+		.nav {
+			background-color: rgba(0, 0, 0, 0.3);
+			overflow: hidden;
+			.nav-item {
+				overflow: hidden;
+				.active {
+					background-color: #20a8d8;
+					span, .sidebar-item-icon {
+						color: #fff !important;
+					}
+				}
+				.nav-link {
+					overflow: hidden;
+					color: #fff;
+					height: 50px;
+					.sidebar-item-icon {
+						color: #464b50;
+					}
+				}
+			}
+		}
+		.nav:hover {
+			.nav-item {
+				.active {
+					background-color: #24b3e7;
+					.sidebar-item-icon {
+						color: #464b50 !important;
+					}
+				}
+				.nav-link {
+					background-color: rgba(0, 0, 0, 0.4);
+					.sidebar-item-icon {
+						color: #464b50;
+					}
+				}
+			}
+		}
+	}
+}
+
+.sidebar-brand-image, .sidebar-brand-image {
+	display: inline-flex;
+	align-items: center;
+	justify-content: center;
+	height: 55px;
+	width: 100%;
+	padding: 0;
+	margin: 0;
+}
