+++ src/styles/constants/colors.scss
@@ -1,6 +1,7 @@
 :root {
 	--font-size-base: .875rem;
 	--white: #FFFFFF;
+	--white-night: #DBEEFF;
 	--gray: #F3F4FA;
 	--dark-gray: #DDE1F8;
 	--black: #0E1426;
