+++ src/styles/constants/dark-mode.scss
@@ -14,6 +14,7 @@
 	--bg-color-dark: var(--black);
 	--border-color: var(--lightest-black);
 
-	--text-color: var(--white);
+	--text-color: var(--white-night);
+	--reversed-text-color: var(--dark-black);
 	--highligted-text-color: var(--violet);
 }
