+++ src/styles/constants/light-mode.scss
@@ -15,5 +15,6 @@ html[data-theme="theme-light"] {
 	--border-color: var(--dark-gray);
 
 	--text-color: var(--dark-black);
+	--reversed-text-color: var(--white);
 	--highligted-text-color: var(--dark-blue);
 }
